requirements.txt
Required Packages:

requests 2.25.1
beautifulsoup4 4.9.3
openpyxl 3.0.6
python-crontab 2.5.1
python-dotenv 0.15.0
mysql-connector-python 8.0.23
Flask 1.1.2
Flask-Cors 3.0.10
mailjet-rest 1.3.4
matplotlib 3.3.2
